requirements.txt
coverage==3.5
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-cors-headers==1.1.0
django-easy-pdf==0.1.0
django-filter==0.11.0
django-postgrespool==0.3.0
django-rest-swagger==0.3.4
django-toolbelt==0.0.1
djangorestframework==3.2.4
drf-nested-routers==0.10.0
gunicorn==19.3.0
html5lib==0.9999999
httplib2==0.7.6
Markdown==2.6.2
nose==1.0.0
Pillow==3.0.0
psycopg2==2.6.1
PyPDF2==1.19
PyYAML==3.11
six==1.10.0
SQLAlchemy==1.0.9
static3==0.6.1
wheel==0.24.0
xhtml2pdf==0.0.6
